start openwhyd server and init db from approval test suite
