+++ .eslintrc.json
@@ -45,6 +45,7 @@
   ],
   "env": {
     "browser": true,
+    "mocha": true,
     "jest": true,
     "node": true,
     "es6": true
