+++ test/approval.tests.js
@@ -1,46 +1,3 @@
-// Run with:              $ npm run test-approval
-// Update snapshots with: $ npm run test-approval -- --update-snapshots
-
-const test = require('ava');
-const { promisify } = require('util');
-const {
-  readMongoDocuments,
-  insertTestData,
-  getCleanedPageBody,
-  startOpenwhydServer,
-} = require('./approval-tests-helpers');
-
-const {
-  START_WITH_ENV_FILE,
-  PORT, // Note: if PORT is not provided, approval-tests-helpers will start Openwhyd's server programmatically, using START_WITH_ENV_FILE
-  DONT_KILL,
-} = process.env;
-
-const MONGODB_URL =
-  process.env.MONGODB_URL || 'mongodb://localhost:27117/openwhyd_test';
-
-test.before(async (t) => {
-  const testDataCollections = {
-    user: await readMongoDocuments(__dirname + '/approval.users.json.js'),
-    post: await readMongoDocuments(__dirname + '/approval.posts.json.js'),
-  };
-  await insertTestData(MONGODB_URL, testDataCollections);
-
-  t.context.serverProcess = await startOpenwhydServer({
-    startWithEnv: START_WITH_ENV_FILE,
-    port: PORT,
-  });
-  t.context.openwhyd = require('./api-client');
-  t.context.getUser = (id) =>
-    testDataCollections.user.find(({ _id }) => id === _id.toString());
-});
-
-test.after((t) => {
-  if (t.context.serverProcess?.kill && !DONT_KILL) {
-    t.context.serverProcess.kill('SIGINT');
-  }
-});
-
 const personas = [
   { label: 'Visitor', userId: undefined },
   { label: 'User: Adrien', userId: '4d94501d1f78ac091dbc9b4d' },
